lab2/user_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
